+++ lib/test/clojure_lsp/feature/paredit_test.clj
@@ -8,9 +8,9 @@
   [paredit-fn code]
   (let [{{row :row col :col} :position zloc :zloc} (h/load-code-into-zloc-and-position code)]
     (if-let [transformations (paredit-fn h/default-uri zloc row col)]
-      (let [{{actual-row :row actual-col :col} :range} (:show-document-after-edit transformations)
+      (let [{{actual-row :row actual-col :col actual-end-col :end-col} :range} (:show-document-after-edit transformations)
             result (h/changes->code (-> transformations :changes-by-uri first second) (h/db))]
-        (h/put-cursor-at result actual-row actual-col))
+        (h/put-cursor-at result actual-row actual-col actual-end-col))
       code)))
 
 (deftest forward-slurp-test
@@ -72,4 +72,40 @@
     "[1 [2 [|] 5] 6]" "[1 [2 [|3 4] 5] 6]"
     "|1" "|1"
     "|" "|"
+    #_()))
+
+(deftest forward-test
+  (are [expected code] (= expected (pareditfy f.paredit/forward code))
+    "(baz [:foo| :bar])" "(baz [|:foo :bar])"
+    "(baz [:foo :bar]|)" "(baz |[:foo :bar])"
+    "(let [foo-bar| (+ 1 2)])" "(let [fo|o-bar (+ 1 2)])"
+    "[1 [2 [] 5]| 6]" "[1 |[2 [] 5] 6]"
+    "{:foo {:bar :b}|}" "{:foo |{:bar :b}}"
+    "(-> foo (+ 1)  (+ 3)| str)" "(-> foo (+ 1) | (+ 3) str)"
+    #_()))
+
+(deftest backward-test
+  (are [expected code] (= expected (pareditfy f.paredit/backward code))
+    "(let [|foo-bar (+ 1 2)])" "(let [foo-b|ar (+ 1 2)])"
+    "[1 |[2 [] 5] 6]" "[1 [2 [] 5]| 6]"
+    "{:foo |{:bar :b}}" "{:foo {:bar :b}|}"
+    "(-> foo |(+ 1)  (+ 3) str)" "(-> foo (+ 1) | (+ 3) str)"
+    #_()))
+
+(deftest forward-select-test
+  (are [expected code] (= expected (pareditfy f.paredit/forward-select code))
+    "(baz [|:foo| :bar])" "(baz [|:foo :bar])"
+    "(baz |[:foo :bar]|)" "(baz |[:foo :bar])"
+    "(let [fo|o-bar| (+ 1 2)])" "(let [fo|o-bar (+ 1 2)])"
+    "[1 |[2 [] 5]| 6]" "[1 |[2 [] 5] 6]"
+    "{:foo |{:bar :b}|}" "{:foo |{:bar :b}}"
+    "(-> foo (+ 1) | (+ 3)| str)" "(-> foo (+ 1) | (+ 3) str)"
+    #_()))
+
+(deftest backward-select-test
+  (are [expected code] (= expected (pareditfy f.paredit/backward-select code))
+    "(let [|foo-b|ar (+ 1 2)])" "(let [foo-b|ar (+ 1 2)])"
+    "[1 |[2 [] 5]| 6]" "[1 [2 [] 5]| 6]"
+    "{:foo |{:bar :b}|}" "{:foo {:bar :b}|}"
+    "(-> foo |(+ 1) | (+ 3) str)" "(-> foo (+ 1) | (+ 3) str)"
     #_()))