+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - General
   - Bump rewrite-clj to `2cf05893ed7af8c0cfe569077ceb12a45738cf2a`. #1984
+- Editor
+  - Add `forward`, `forward-select`, `backward` and `backward-select` paredit actions.
 
 ## 2025.03.07-17.42.36